% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04p.R
\name{graph_sum_n}
\alias{graph_sum_n}
\title{Graph of Sum of n numerics}
\usage{
graph_sum_n(df)
}
\arguments{
\item{a}{data frame or tibble}
}
\value{
graph of outputs of sum_n against the value of n
}
\description{
#Takes a tibble or data frame and using sum_n calculates the nth sum.
}
\examples{
test <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
graph_sum_n(test)
error_df1 <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3.5
)
error_df2 <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,0
)
error_df3 <-tribble(
~x, ~y, ~z, ~n,
2,4,3,-1
)

\dontrun{ graph_sum_n(error_df1), graph_sum_n(error_df2), graph_sum_n(error_df3)}
}
