% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pChop18-package.R
\docType{package}
\name{hw04pChop18-package}
\alias{hw04pChop18}
\alias{hw04pChop18-package}
\title{hw04pChop18: hw04pChop18}
\description{
The package consists of two functions. Function 1 finds the sum of a vector of three values up to the nth term.Function 2 graphs the output of function 1 against the nth value.
}
\author{
\strong{Maintainer}: Christopher Hopkins \email{ch3222a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
