% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04p.R
\name{sum_n}
\alias{sum_n}
\title{Sum of n numerics}
\usage{
sum_n(v, n)
}
\arguments{
\item{vector}{of three numerics}

\item{integer}{}
}
\value{
a sum of numerics
}
\description{
recursive function to find the nth sum of a sequence of values.
The first three values must be defined and the nth term must be a integer greater than 0.
}
\examples{
sum_n(c( 2, 4, 3), 5)
\dontrun{ sum_n(c( 2, 4), 5), sum_n(c( 2, 4, 3), 5.5), sum_n(c( 2, 4, 3), -5)}
}
