% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04p.R
\name{graph_sum_n}
\alias{graph_sum_n}
\title{Graph of Sum of n numerics}
\usage{
graph_sum_n(df)
}
\arguments{
\item{data}{frame or tibble}
}
\value{
graph of outputs of sum_n against the value of n
}
\description{
#Takes a tibble or data frame and using sum_n calculates the nth sum.
}
\examples{
test <- tibble::tibble(
x = 2,
y = 4,
z = 3,
n = 3:12
)

graph_sum_n(test)

error_df1 <- tibble::tibble(
x = 2,
y = 3,
n = 4
)

error_df2 <- tibble::tibble(
x = 2,
y = 4,
z = 3,
n = 0
)

error_df3 <- tibble::tibble(
x = 2,
y = 4,
z = 3,
n = -1
)

error_df4 <- tibble::tibble(
x = 2,
y = 4,
z = 3,
n = 6.5
)
\dontrun{ graph_sum_n(error_df1),
graph_sum_n(error_df2),
graph_sum_n(error_df3),
graph_sum_n(error_df4)}
}
